\name{cora_view4}
\alias{cora_view4}
\docType{data}
\title{
The fourth view of Cora data set.
}
\description{
The fourth view(outbound view) of Cora data set.  Cora data set is a multi-view data set of machine learning papers with 4 views, a sample size of nearly 3000 and a number of features of 1500, with a number of clusters of K=4.
}
\usage{data("cora_view4")}
\format{
  The format is:
 num [1:2708, 1:2708] 0 0 0 0 0 0 0 0 1 0 ...
}
\details{
The fourth view of Cora data set.
}
\source{
http://www.cs.umd.edu/projects/linqs/projects/lbc/
}
\references{
http://www.cs.umd.edu/projects/linqs/projects/lbc/
}
\examples{
data(cora_view4); str(cora_view4) 
}
\keyword{data sets}
