\name{labelcora}
\alias{labelcora}
\docType{data}
\title{
True clustering labels for Cora data set.
}
\description{
True clustering labels for the Cora dataset, which can be applied to 4 views.
}
\usage{data("labelcora")}
\format{
  The format is:
 chr [1:2708] "1" "2" "3" "3" "4" "4" "5" "1" "1" "5" "1" "6" "4" "7" ...
}
\details{
True clustering labels for the Cora dataset, which can be applied to 4 views.
}
\source{
http://www.cs.umd.edu/projects/linqs/projects/lbc/
}
\references{
http://www.cs.umd.edu/projects/linqs/projects/lbc/
}
\examples{
data(labelcora)
}
\keyword{datasets}
