\name{QCM}
\alias{QCM}
\docType{data}
\title{
The QCM data set with K=5.
}
\description{
The QCM data set with K=5.
}
\usage{data("QCM")}
\format{
  A data frame with 125 observations on the following 15 variables.
  \describe{
    \item{\code{V1}}{a numeric vector}
    \item{\code{V2}}{a numeric vector}
    \item{\code{V3}}{a numeric vector}
    \item{\code{V4}}{a numeric vector}
    \item{\code{V5}}{a numeric vector}
    \item{\code{V6}}{a numeric vector}
    \item{\code{V7}}{a numeric vector}
    \item{\code{V8}}{a numeric vector}
    \item{\code{V9}}{a numeric vector}
    \item{\code{V10}}{a numeric vector}
    \item{\code{V11}}{a numeric vector}
    \item{\code{V12}}{a numeric vector}
    \item{\code{V13}}{a numeric vector}
    \item{\code{V14}}{a numeric vector}
    \item{\code{V15}}{a numeric vector}
  }
}
\details{
The QCM data set with K=5.
}
\source{
https://www.sciencedirect.com/science/article/pii/S2215098619303337.
}
\references{
M. F. Adak, P. Lieberzeit, P. Jarujamrus, and N. Yumusak. Classification of alcohols obtained by qcm sensors with different characteristics using abc based neural network. Engineering Science and Technology, an International Journal, 23(3):463–469, 2020. ISSN 2215-0986. doi: https://doi.
org/10.1016/j.jestch.2019.06.011. URL https://www.sciencedirect.com/science/article/pii/S2215098619303337.
}
\examples{
data(QCM)
## maybe str(QCM) ; plot(QCM) ...
}
\keyword{datasets}
