\name{seed}
\alias{seed}
\docType{data}
\title{
A single-view data set named seed.
}
\description{
A single-view data set named seed.
}
\usage{data("seed")}
\format{
  A data frame with 210 observations on the following 8 variables.
  \describe{
    \item{\code{V1}}{a numeric vector}
    \item{\code{V2}}{a numeric vector}
    \item{\code{V3}}{a numeric vector}
    \item{\code{V4}}{a numeric vector}
    \item{\code{V5}}{a numeric vector}
    \item{\code{V6}}{a numeric vector}
    \item{\code{V7}}{a numeric vector}
    \item{\code{V8}}{a numeric vector}
  }
}
\details{
A single-view data set named seed.
}
\source{
http://archive.ics.uci.edu/ml/datasets/seeds
}
\references{
http://archive.ics.uci.edu/ml/datasets/seeds
}
\examples{
data(seed)
## maybe str(seed) ; plot(seed) ...
}
\keyword{datasets}
