\name{seedX}
\alias{seedX}
\docType{data}
\title{
A single-view data set named Seeds.
}
\description{
The Seeds data set holds data on the area, circumference, compaction, seed length, seed width, asymmetry factor, length of the ventral groove of the seed and category data for different varieties of wheat seeds. The data set contains a total of 210 records, 7 features, and one label, which is divided into 3 categories.
}
\usage{data("seedX")}
\format{
  The format is:
 num [1:210, 1:8] 15.3 14.9 14.3 13.8 16.1 ...
}
\details{
A single-view data set named seed.
}
\source{
http://archive.ics.uci.edu/ml/datasets/seeds
}
\references{
http://archive.ics.uci.edu/ml/datasets/seeds
}
\examples{
data(seedX); str(seedX) 
}
\keyword{datasets}
