\name{sobar}
\alias{sobar}
\docType{data}
\title{
A single-view data set named sobar.
}
\description{
A single-view data set named sobar.
}
\usage{data("sobar")}
\format{
  A data frame with 72 observations on the following 19 variables.
  \describe{
    \item{\code{behavior_sexualRisk}}{a numeric vector}
    \item{\code{behavior_eating}}{a numeric vector}
    \item{\code{behavior_personalHygine}}{a numeric vector}
    \item{\code{intention_aggregation}}{a numeric vector}
    \item{\code{intention_commitment}}{a numeric vector}
    \item{\code{attitude_consistency}}{a numeric vector}
    \item{\code{attitude_spontaneity}}{a numeric vector}
    \item{\code{norm_significantPerson}}{a numeric vector}
    \item{\code{norm_fulfillment}}{a numeric vector}
    \item{\code{perception_vulnerability}}{a numeric vector}
    \item{\code{perception_severity}}{a numeric vector}
    \item{\code{motivation_strength}}{a numeric vector}
    \item{\code{motivation_willingness}}{a numeric vector}
    \item{\code{socialSupport_emotionality}}{a numeric vector}
    \item{\code{socialSupport_appreciation}}{a numeric vector}
    \item{\code{socialSupport_instrumental}}{a numeric vector}
    \item{\code{empowerment_knowledge}}{a numeric vector}
    \item{\code{empowerment_abilities}}{a numeric vector}
    \item{\code{empowerment_desires}}{a numeric vector}
  }
}
\details{
A single-view data set named sobar.
}
\source{
http://archive.ics.uci.edu/ml/datasets/Cervical+Cancer+Behavior+Risk
}
\references{
http://archive.ics.uci.edu/ml/datasets/Cervical+Cancer+Behavior+Risk
}
\examples{
data(sobar)
## maybe str(sobar) ; plot(sobar) ...
}
\keyword{datasets}
