\name{turelabel}
\alias{turelabel}
\docType{data}
\title{
Ture label of Movie data set.
}
\description{
Ture label of Movie data set. You can use it to calculate the accuracy of the clustering results.
}
\usage{data("turelabel")}
\format{
  A data frame with 617 observations on the following variable.
  \describe{
    \item{\code{V1}}{a numeric vector}
  }
}
\details{
Ture label of Movie data set.
}
\source{
 https://lig-membres.imag.fr/grimal/data.html.
}
\references{
C. Grimal. the multi-view movie data set. 2010. URL https://lig-membres.imag.fr/grimal/data.html.
}
\examples{
data(turelabel)
## maybe str(turelabel) ; plot(turelabel) ...
}
\keyword{data sets}
