\name{QCM}
\alias{QCM}
\docType{data}
\title{
The QCM data set with K=5.
}
\description{
Five different QCM gas sensors were used and five different gas measurements were made for each sensor (1-octanol, 1-propanol, 2-butanol, 2-propanol and 1-isobutanol).
}
\usage{data("QCM")}
\format{
  The format is:
 num [1:125, 1:15] -10.06 -9.69 -12.07 -14.21 -16.57 ...
}
\details{
The QCM data set with K=5.
}
\source{
https://www.sciencedirect.com/science/article/pii/S2215098619303337.
}
\references{
M. F. Adak, P. Lieberzeit, P. Jarujamrus, and N. Yumusak. Classification of alcohols obtained by qcm sensors with different characteristics using abc based neural network. Engineering Science and Technology, an International Journal, 23(3):463–469, 2020. ISSN 2215-0986. doi: https://doi.
org/10.1016/j.jestch.2019.06.011. URL https://www.sciencedirect.com/science/article/pii/S2215098619303337.
}
\examples{
data(QCM); str(QCM) 
}
\keyword{data sets}
