\name{sobar}
\alias{sobar}
\docType{data}
\title{
A single-view data set named Sobar.
}
\description{
A single-view data set named Sobar. The Sobar data set is a behavioural risk dataset for cervical cancer, which has a number of clusters of 2.
}
\usage{data("sobar")}
\format{
  The format is:
 num [1:72, 1:20] 10 10 10 10 8 10 10 8 10 7 ...
}
\details{
A single-view data set named sobar.
}
\source{
http://archive.ics.uci.edu/ml/datasets/Cervical+Cancer+Behavior+Risk
}
\references{
http://archive.ics.uci.edu/ml/datasets/Cervical+Cancer+Behavior+Risk
}
\examples{
data(sobar); str(sobar) 
}
\keyword{datasets}
